pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "A tool to transcribe audio files with speaker diarization."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.1.2",
    "torchaudio==2.1.2",
    "pyannote.audio==3.1.1",
    "numpy==1.26.4",
    "whisperx",
    "hf-transfer",
    "safetensors",
    "tqdm",
]

[project.scripts]
transcribe = "transcribe:main"
